move the derived class to its own file
