+++ src/dft/backends/mklcpu/commit.cpp
@@ -32,7 +32,7 @@
 
 #include "oneapi/mkl/dft/detail/commit_impl.hpp"
 
-#include "dft/backends/mklcpu/mklcpu_helpers.hpp"
+#include "dft/backends/mklcpu/commit_derived_impl.hpp"
 #include "mkl_service.h"
 #include "mkl_dfti.h"
