+++ src/dft/backends/mklcpu/mklcpu_helpers.hpp
@@ -173,46 +173,7 @@ inline constexpr int to_mklcpu<dft::detail::config_param::PACKED_FORMAT>(
     }
 }
 
-// this is used for indexing bidirectional_handle
-enum DIR { fwd=0, bwd=1 };
-
-template <dft::detail::precision prec, dft::detail::domain dom>
-class commit_derived_impl : public dft::detail::commit_impl<prec, dom> {
-private:
-    static constexpr DFTI_CONFIG_VALUE mklcpu_prec = to_mklcpu(prec);
-    static constexpr DFTI_CONFIG_VALUE mklcpu_dom = to_mklcpu(dom);
-    using mklcpu_desc_t = DFTI_DESCRIPTOR_HANDLE;
-
-public:
-    commit_derived_impl(sycl::queue queue, const dft::detail::dft_values<prec, dom>& config_values);
-
-    virtual void commit(const dft::detail::dft_values<prec, dom>& config_values) override;
-
-    virtual void* get_handle() noexcept override;
-
-    virtual ~commit_derived_impl() override;
-
-    virtual sycl::buffer<std::array<mklcpu_desc_t, 2>, 1> get_handle_buffer() noexcept;
-
-private:
-    // bidirectional_handle[0] is the forward handle, bidirectional_handle[1] is the backward handle
-    std::array<mklcpu_desc_t, 2> bidirection_handle{ nullptr, nullptr };
-    sycl::buffer<std::array<mklcpu_desc_t, 2>, 1> bidirection_buffer{ &bidirection_handle,
-                                                                    sycl::range<1>{ 2 } };
-
-    template <typename... Args>
-    void set_value_item(mklcpu_desc_t hand, enum DFTI_CONFIG_PARAM name, Args... args);
-
-    void set_value(mklcpu_desc_t* descHandle, const dft::detail::dft_values<prec, dom>& config);
-};
-
 using mklcpu_desc_t = DFTI_DESCRIPTOR_HANDLE;
-
-template <dft::detail::precision prec, dft::detail::domain dom>
-using commit_t = dft::detail::commit_impl<prec, dom>;
-
-template <dft::detail::precision prec, dft::detail::domain dom>
-using commit_derived_t = detail::commit_derived_impl<prec, dom>;
 
 } // namespace detail
 } // namespace mklcpu
